Preparing next upgrade 1.10.1
git-svn-id: https://bobcat.svn.sourceforge.net/svnroot/bobcat/trunk@107 95c995b8-3b1c-0410-acf1-adacd5d40d98

+++ bobcat/arg/operatorindex.cc
@@ -1,6 +1,6 @@
 #include "arg.ih"
 
-char const *Arg::operator[](unsigned idx) const
+char const *Arg::operator[](size_t idx) const
 {
     return
         (idx >= nArgs()) ?
