+++ bobcat/arg/option6.cc
@@ -1,6 +1,6 @@
 #include "arg.ih"
 
-unsigned Arg::option(unsigned *idx, string *value, char const *longOpt) const
+size_t Arg::option(size_t *idx, string *value, char const *longOpt) const
 {
     SSVMapIterator it = d_longOptv.find(longOpt);
     if (it == d_longOptv.end())
