+++ bobcat/arg/versionhelp.cc
@@ -1,7 +1,7 @@
 #include "arg.ih"
 
 void Arg::versionHelp(void (*usage)(string const &progname), 
-        char const *version, unsigned minArgs, int helpFlag, 
+        char const *version, size_t minArgs, int helpFlag, 
         int versionFlag) const throw (int)
 {
     if (option(versionFlag) && !option(helpFlag))
