+++ bobcat/hostent/countaliases.cc
@@ -1,8 +1,8 @@
 #include "hostent.ih"
 
-unsigned Hostent::countAliases(char const * const *alias) const
+size_t Hostent::countAliases(char const * const *alias) const
 {
-    unsigned idx = 1;   // counts the final 0-ptr.
+    size_t idx = 1;   // counts the final 0-ptr.
 
     while (*alias++)
         idx++;
