+++ bobcat/log/level.cc
@@ -1,6 +1,6 @@
 #include "log.ih"
 
-ostream &Log::level(unsigned activeLevel)
+ostream &Log::level(size_t activeLevel)
 {
     d_activeLevel = activeLevel;
     setActive(d_level <= d_activeLevel);
