+++ bobcat/ofdstreambuf/open.cc
@@ -1,6 +1,6 @@
 #include "ofdstreambuf.ih"
 
-void OFdStreambuf::open(int fd, Mode mode, unsigned n)
+void OFdStreambuf::open(int fd, Mode mode, size_t n)
 {
     cleanup(mode);
