+++ bobcat/xpointer/data.cc
@@ -1,4 +1,4 @@
 #include "xpointer.ih"
 
-unsigned Xpointer::s_counter = 0;
+size_t Xpointer::s_counter = 0;
 Display *Xpointer::s_theDisplay;
