+++ bobcat/xpointer/get.cc
@@ -5,7 +5,7 @@ bool Xpointer::get(int *x, int *y) const
     Window rootret;
     Window child;
     Window root = DefaultRootWindow(s_theDisplay);
-    unsigned keys;
+    size_t keys;
     int dum_x;
     int dum_y;
